fix: counts & save button align
