+++ src/taxonomy/tree-table/CreateRow.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { IntlProvider } from '@edx/frontend-platform/i18n';
 import { fireEvent, render, screen } from '@testing-library/react';
 
-import { CreateRow } from './CreateRow';
+import CreateRow from './CreateRow';
 
 const wrapper = ({ children }: { children: React.ReactNode }) => (
   <IntlProvider locale="en" messages={{}}>{children}</IntlProvider>
